move fixture_probabilities back to api_utils to avoid circular dependency
