+++ airsenal/framework/utils.py
@@ -130,7 +130,7 @@ def get_next_gameweek(season=CURRENT_SEASON, dbsession=None):
 NEXT_GAMEWEEK = get_next_gameweek()
 
 
-from .bpl_interface import get_fitted_team_model
+
 
 
 def get_previous_season(season):
@@ -1044,46 +1044,6 @@ def get_latest_fixture_tag(season=CURRENT_SEASON, dbsession=None):
         dbsession = session
     rows = dbsession.query(Fixture).filter_by(season=season).all()
     return rows[-1].tag
-
-
-def fixture_probabilities(gameweek, season=CURRENT_SEASON, dbsession=None):
-    """
-    Returns probabilities for all fixtures in a given gameweek and season, as a data frame with a row
-    for each fixture and columns being fixture_id, home_team, away_team, home_win_probability,
-    draw_probability, away_win_probability.
-    """
-    model_team = get_fitted_team_model(season, dbsession)
-    fixture_probabilities_list = []
-    fixture_id_list = []
-    for fixture in get_fixtures_for_gameweek(
-        gameweek, season=season, dbsession=dbsession
-    ):
-        probabilities = model_team.overall_probabilities(
-            fixture.home_team, fixture.away_team
-        )
-        fixture_probabilities_list.append(
-            [
-                fixture.fixture_id,
-                fixture.home_team,
-                fixture.away_team,
-                probabilities[0],
-                probabilities[1],
-                probabilities[2],
-            ]
-        )
-        fixture_id_list.append(fixture.fixture_id)
-    return pd.DataFrame(
-        fixture_probabilities_list,
-        columns=[
-            "fixture_id",
-            "home_team",
-            "away_team",
-            "home_win_probability",
-            "draw_probability",
-            "away_win_probability",
-        ],
-        index=fixture_id_list,
-    )
 
 
 def find_fixture(
